Move tests into provider module.
