+++ provider/src/provider.rs
@@ -381,3 +381,220 @@ impl<C: Client> Poller<C> {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::lambda::{
+        InitializationError, InvocationError, InvocationEvent, InvocationEventBuilder,
+        InvocationResponse,
+    };
+    use std::cell::RefCell;
+
+    use crate::tests_common::*;
+
+    /// Represents the event kind to return.
+    enum EventKind {
+        /// No event.
+        None,
+        /// An invocation event.
+        Event(InvocationEvent),
+        /// An error.
+        Error,
+    }
+
+    /// Represents a mock Lambda runtime client that returns a single event.
+    struct MockClient {
+        event_kind: EventKind,
+        initialization_error: RefCell<Option<InitializationError>>,
+        invocation_error: RefCell<Option<InvocationError>>,
+        invocation_response: RefCell<Option<InvocationResponse>>,
+        shutdown_map: ShutdownMap,
+    }
+
+    impl MockClient {
+        /// Returns a new `MockClient`.
+        fn new(event_kind: EventKind, shutdown_map: ShutdownMap) -> Self {
+            Self {
+                event_kind,
+                initialization_error: RefCell::new(None),
+                invocation_error: RefCell::new(None),
+                invocation_response: RefCell::new(None),
+                shutdown_map,
+            }
+        }
+    }
+
+    impl Client for MockClient {
+        /// Returns the next AWS Lambda invocation event.
+        fn next_invocation_event(&self) -> anyhow::Result<Option<InvocationEvent>> {
+            // Shutdown after one event.
+            self.shutdown_map.put(MODULE_ID, true)?;
+
+            match &self.event_kind {
+                EventKind::None => Ok(None),
+                EventKind::Event(event) => Ok(Some(event.clone())),
+                EventKind::Error => Err(anyhow!(ERROR_MESSAGE)),
+            }
+        }
+
+        /// Sends an invocation error to the AWS Lambda runtime.
+        fn send_invocation_error(&self, error: InvocationError) -> anyhow::Result<()> {
+            *self.invocation_error.borrow_mut() = Some(error);
+            Ok(())
+        }
+
+        /// Sends an invocation error to the AWS Lambda runtime.
+        fn send_invocation_response(&self, response: InvocationResponse) -> anyhow::Result<()> {
+            *self.invocation_response.borrow_mut() = Some(response);
+            Ok(())
+        }
+
+        /// Sends an initialization error to the AWS Lambda runtime.
+        fn send_initialization_error(&self, error: InitializationError) -> anyhow::Result<()> {
+            *self.initialization_error.borrow_mut() = Some(error);
+            Ok(())
+        }
+    }
+
+    impl InvocationEvent {
+        /// Returns an `InvocationEvent` with no request ID.
+        fn no_request_id() -> Self {
+            InvocationEventBuilder::new(EVENT_BODY.to_vec()).build()
+        }
+
+        /// Return an `InvocationEvent` with a request ID.
+        fn with_request_id() -> Self {
+            let mut builder = InvocationEventBuilder::new(EVENT_BODY.to_vec());
+            builder = builder.request_id(REQUEST_ID);
+            builder.build()
+        }
+    }
+
+    /// Returns a test event dispatcher.
+    fn dispatcher() -> impl InvocationEventDispatcher {
+        let response = codec::Response {
+            body: RESPONSE_BODY.to_vec(),
+        };
+        let host_dispatcher: HostDispatcher =
+            Arc::new(RwLock::new(Box::new(MockHostDispatcher::new(response))));
+        EventDispatcher::new(host_dispatcher)
+    }
+
+    /// Returns a test event dispatcher that errors.
+    fn error_dispatcher() -> impl InvocationEventDispatcher {
+        let host_dispatcher: HostDispatcher =
+            Arc::new(RwLock::new(Box::new(ErrorHostDispatcher::new())));
+        EventDispatcher::new(host_dispatcher)
+    }
+
+    /// Returns a mock poller.
+    fn mock_poller(event_kind: EventKind) -> Poller<MockClient> {
+        let shutdown_map = ShutdownMap::new();
+        Poller::new(
+            MODULE_ID,
+            MockClient::new(event_kind, shutdown_map.clone()),
+            shutdown_map,
+        )
+    }
+
+    /// Tests that receiving an empty event sends no response or error.
+    #[test]
+    fn event_kind_none() {
+        let poller = mock_poller(EventKind::None);
+        let result = poller.shutdown_map.put(MODULE_ID, false);
+        assert!(result.is_ok());
+
+        poller.run(dispatcher());
+
+        assert!(poller.client.initialization_error.borrow().is_none());
+        assert!(poller.client.invocation_response.borrow().is_none());
+        assert!(poller.client.invocation_error.borrow().is_none());
+    }
+
+    /// Tests that receiving an event without a request ID sends no response or error.
+    #[test]
+    fn event_kind_event_no_request_id() {
+        let poller = mock_poller(EventKind::Event(InvocationEvent::no_request_id()));
+        let result = poller.shutdown_map.put(MODULE_ID, false);
+        assert!(result.is_ok());
+
+        poller.run(dispatcher());
+
+        assert!(poller.client.initialization_error.borrow().is_none());
+        assert!(poller.client.invocation_response.borrow().is_none());
+        assert!(poller.client.invocation_error.borrow().is_none());
+    }
+
+    /// Tests that receiving an event with a request ID sends a response.
+    #[test]
+    fn event_kind_event_with_request_id() {
+        let poller = mock_poller(EventKind::Event(InvocationEvent::with_request_id()));
+        let result = poller.shutdown_map.put(MODULE_ID, false);
+        assert!(result.is_ok());
+
+        poller.run(dispatcher());
+
+        assert!(poller.client.initialization_error.borrow().is_none());
+        assert!(poller.client.invocation_response.borrow().is_some());
+        assert_eq!(
+            REQUEST_ID,
+            poller
+                .client
+                .invocation_response
+                .borrow()
+                .as_ref()
+                .unwrap()
+                .request_id()
+        );
+        assert_eq!(
+            RESPONSE_BODY,
+            poller
+                .client
+                .invocation_response
+                .borrow()
+                .as_ref()
+                .unwrap()
+                .body()
+        );
+        assert!(poller.client.invocation_error.borrow().is_none());
+    }
+
+    /// Tests that receiving an event with a request ID and dispatching with an error sends an error.
+    #[test]
+    fn event_kind_event_with_request_id_error_dispatcher() {
+        let poller = mock_poller(EventKind::Event(InvocationEvent::with_request_id()));
+        let result = poller.shutdown_map.put(MODULE_ID, false);
+        assert!(result.is_ok());
+
+        poller.run(error_dispatcher());
+
+        assert!(poller.client.initialization_error.borrow().is_none());
+        assert!(poller.client.invocation_response.borrow().is_none());
+        assert!(poller.client.invocation_error.borrow().is_some());
+        assert_eq!(
+            REQUEST_ID,
+            poller
+                .client
+                .invocation_error
+                .borrow()
+                .as_ref()
+                .unwrap()
+                .request_id()
+        );
+    }
+
+    /// Tests that receiving an error sends no response or error.
+    #[test]
+    fn event_kind_error() {
+        let poller = mock_poller(EventKind::Error);
+        let result = poller.shutdown_map.put(MODULE_ID, false);
+        assert!(result.is_ok());
+
+        poller.run(dispatcher());
+
+        assert!(poller.client.initialization_error.borrow().is_none());
+        assert!(poller.client.invocation_response.borrow().is_none());
+        assert!(poller.client.invocation_error.borrow().is_none());
+    }
+}
